refactor: create sidebar-selectors file to keep related code together
