+++ packages/insomnia-app/app/ui/redux/__tests__/sidebar-selectors.test.js
@@ -1,5 +1,9 @@
 import * as models from '../../../models';
-import { shouldIgnoreChildrenOf, shouldShowInSidebar, sortByMetaKeyOrId } from '../helpers';
+import {
+  shouldIgnoreChildrenOf,
+  shouldShowInSidebar,
+  sortByMetaKeyOrId,
+} from '../sidebar-selectors';
 import { difference } from 'lodash';
 
 describe('shouldShowInSidebar', () => {
